fix: rename registry.py to fix class finding when pickling
