+++ edsnlp/__init__.py
@@ -10,7 +10,7 @@ from spacy import pipeline as _spacy_pipeline  # noqa: F401
 from . import extensions
 from . import patch_spacy
 from .core.pipeline import Pipeline, blank, load
-from .core.registry import registry
+from .core.registries import registry
 import edsnlp.data  # noqa: F401
 import edsnlp.pipes
